requirements.txt

# Python dependencies for Analysis & Machine Learning
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
google-cloud-aiplatform>=1.35.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.28.0
beautifulsoup4>=4.11.0 # For basic parsing if needed
jupyter>=1.0.0
